requirements.txt
Flask==3.1.1
requests==2.32.3
bs4==0.0.2
ollama==0.4.8
langchain-community==0.3.24
transformers==4.51.3
torch==2.7.0
flask-cors==6.0.0
langchain-core==0.3.60
langchain-ollama==0.3.3
sentence-transformers==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
boxsdk==3.14.0
chardet==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
faiss-cpu==1.11.0
psutil
email_validator
pypdf
